Add curl command
The biggest flaw in using this toolkit is that queries are stored in
python format, which only works for python. If you want to share a
query, you have to manually extract the body, ensure it's JSON-compliant
(which it likely won't be due to some commas), and then build a CURL
command from scratch.

No more! Now a command handles this for you (could have saved me a
cumulutive hour by now if I did this a year ago...)